Farm_AI/ai_service/flask-backend/requirements.txt
Flask
flask_cors
numpy
pandas
requests
scikit-learn
Pillow
torch
torchvision
google-generativeai
python-dotenv
